# Conflicts:
#	src/main/java/com/csidigital/rh/management/controller/InterviewController.java
#	src/main/java/com/csidigital/rh/management/service/impl/EvaluationImpl.java
#	src/main/java/com/csidigital/rh/management/service/impl/InterviewImpl.java
#	src/main/java/com/csidigital/rh/management/service/impl/QuestionTypeImpl.java
#	src/main/java/com/csidigital/rh/shared/dto/request/EvaluationRequest.java
#	src/main/java/com/csidigital/rh/shared/dto/response/EvaluationResponse.java

+++ src/main/java/com/csidigital/rh/dao/entity/Evaluation.java
@@ -19,7 +19,7 @@ public class Evaluation{
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @Column(name = "globalAppreciation")
-    private Integer globalAppreciation;
+    private Double globalAppreciation;
     private String evaluationRef;
     private LocalDate evaluationDate;
